+++ src/Raven.CodeAnalysis/BoundTree/BoundBlockExpression.cs
@@ -1,0 +1,12 @@
+namespace Raven.CodeAnalysis;
+
+class BoundBlockExpression : BoundExpression
+{
+    public BoundBlockExpression(IReadOnlyList<BoundExpression> statements)
+        : base(statements.LastOrDefault()?.Type, null, CandidateReason.None)
+    {
+        Statements = statements;
+    }
+
+    public IReadOnlyList<BoundExpression> Statements { get; }
+}
