+++ src/Raven.CodeAnalysis/CompilationOptions.cs
@@ -1,0 +1,16 @@
+namespace Raven.CodeAnalysis;
+
+public class CompilationOptions
+{
+    public CompilationOptions()
+    {
+        OutputKind = OutputKind.ConsoleApplication;
+    }
+
+    public CompilationOptions(OutputKind outputKind)
+    {
+        OutputKind = outputKind;
+    }
+
+    public OutputKind OutputKind { get; }
+}
