+++ src/Raven.CodeAnalysis/NullableAnnotation.cs
@@ -1,0 +1,8 @@
+namespace Raven.CodeAnalysis;
+
+public enum NullableAnnotation
+{
+    None,
+    NotAnnotated,
+    Annotated
+}
