+++ src/Raven.CodeAnalysis/NullableFlowState.cs
@@ -1,0 +1,8 @@
+namespace Raven.CodeAnalysis;
+
+public enum NullableFlowState
+{
+    None,
+    NotNull,
+    MaybeNull
+}
