+++ src/Raven.CodeAnalysis/OutputKind.cs
@@ -1,0 +1,8 @@
+namespace Raven.CodeAnalysis;
+
+public enum OutputKind
+{
+    ConsoleApplication = 0,
+    WindowsApplication = 1,
+    DynamicallyLinkedLibrary = 2
+}
