+++ src/Raven.CodeAnalysis/SemanticModel.cs
@@ -3,7 +3,6 @@ using System.Reflection;
 using System.Reflection.Metadata;
 using System.Runtime.CompilerServices;
 
-using Raven.CodeAnalysis.Symbols;
 using Raven.CodeAnalysis.Syntax;
 
 namespace Raven.CodeAnalysis;
@@ -27,16 +26,6 @@ public partial class SemanticModel
         Compilation = compilation;
         SyntaxTree = syntaxTree;
 
-        /*
-        _keywordTypeSymbols = new Dictionary<string, ITypeSymbol>
-        {
-            { "int", Compilation.GetSpecialType(SpecialType.System_Int32) },
-            { "bool", Compilation.GetSpecialType(SpecialType.System_Boolean) },
-            { "char", Compilation.GetSpecialType(SpecialType.System_Char) },
-            { "string", Compilation.GetSpecialType(SpecialType.System_String) }
-        };
-        */
-
         CreateModel();
     }
 
@@ -45,37 +34,6 @@ public partial class SemanticModel
         // Optional: preload binder for compilation unit to cache imports
         var root = SyntaxTree.GetRoot();
         _ = _binderFactory.GetBinder(root);
-
-        // Diagnostics will be filled lazily as semantic queries come in
-
-        /*
-        foreach (var symbol in _symbols.OfType<SourceMethodSymbol>())
-        {
-            var syntaxRef = symbol.DeclaringSyntaxReferences.First();
-            var syntax = syntaxRef.GetSyntax();
-
-            if (syntax is MethodDeclarationSyntax methodDeclaration)
-            {
-
-            }
-            else if (syntax is CompilationUnitSyntax compilationUnit)
-            {
-                foreach (var import in compilationUnit.Imports)
-                {
-                    var namespaceSymbol = Compilation.GetNamespaceSymbol(import.NamespaceOrType.ToString());
-                    _imports.Add(import.NamespaceOrType, namespaceSymbol);
-                }
-
-                foreach (var member in compilationUnit.Members)
-                {
-                    if (member is GlobalStatementSyntax globalStatement)
-                    {
-                        AnalyzeStatement(symbol, globalStatement.Statement);
-                    }
-                }
-            }
-        }
-        */
     }
 
     public Compilation Compilation { get; }
@@ -84,7 +42,27 @@ public partial class SemanticModel
 
     private DiagnosticBag Diagnostics => _diagnostics;
 
-    public IImmutableList<Diagnostic> GetDiagnostics(CancellationToken cancellationToken = default) => _diagnostics.ToImmutableArray();
+    public IImmutableList<Diagnostic> GetDiagnostics(CancellationToken cancellationToken = default)
+    {
+        EnsureDiagnosticsCollected();
+
+        return _binderFactory
+             .GetAllBinders()
+             .SelectMany(b => b.Diagnostics.ToImmutableArray())
+             .DistinctBy(d => (d.Descriptor.Id, d.Location, d.GetMessage()))
+             .ToImmutableArray();
+    }
+
+    private void EnsureDiagnosticsCollected()
+    {
+        var root = SyntaxTree.GetRoot();
+
+        foreach (var globalStmt in root.DescendantNodes().OfType<GlobalStatementSyntax>())
+        {
+            var binder = _binderFactory.GetBinder(globalStmt);
+            binder.BindStatement(globalStmt.Statement);
+        }
+    }
 
     public SymbolInfo GetSymbolInfo(SyntaxNode node, CancellationToken cancellationToken = default)
     {
@@ -97,22 +75,6 @@ public partial class SemanticModel
         _bindings[node] = info;
         return info;
     }
-
-    /*
-    public SymbolInfo GetSymbolInfo(ExpressionSyntax expr)
-    {
-        var binder = _binderFactory.GetBinder(expr);
-        var bound = binder.BindExpression(expr);
-        return new SymbolInfo(bound.Symbol);
-    }
-
-    public TypeInfo GetTypeInfo(ExpressionSyntax expr)
-    {
-        var binder = _binderFactory.GetBinder(expr);
-        var bound = binder.BindExpression(expr);
-        return new TypeInfo(bound.Type);
-    }
-    */
 
     public ISymbol? GetDeclaredSymbol(SyntaxNode node)
     {
@@ -137,16 +99,6 @@ public partial class SemanticModel
         _bindings[node] = new SymbolInfo(reason, symbols.ToImmutableArray());
     }
 
-    /*
-    public TypeInfo GetTypeInfo(SyntaxNode node, CancellationToken cancellationToken = default)
-    {
-        var binder = _binderFactory.GetBinder(node);
-        var type = binder.BindType(node)?.UnwrapType();
-
-        return new TypeInfo(type, convertedType: null);
-    }
-    */
-
     private ITypeSymbol? InferLiteralType(LiteralExpressionSyntax literal)
     {
         // Example inference logic for literals
@@ -165,108 +117,5 @@ public partial class SemanticModel
         return _symbols
             .OfType<ITypeSymbol>()
             .FirstOrDefault(x => x.ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat) == typeName);
-    }
-}
-
-internal class ErrorTypeSymbol : Symbol, IErrorTypeSymbol
-{
-    public ErrorTypeSymbol(string name, ISymbol containingSymbol, Location[] locations, SyntaxReference[] declaringSyntaxReferences)
-        : base(SymbolKind.ErrorType, name, containingSymbol, null, null, locations, declaringSyntaxReferences)
-    {
-    }
-
-    public ImmutableArray<IMethodSymbol> Constructors => [];
-
-    public IMethodSymbol? StaticConstructor => null;
-
-    public ImmutableArray<ITypeSymbol> TypeArguments => [];
-
-    public ImmutableArray<ITypeParameterSymbol> TypeParameters => [];
-
-    public SpecialType SpecialType => SpecialType.None;
-
-    public bool IsValueType => false;
-
-    public bool IsNamespace => false;
-
-    public bool IsType => false;
-
-    public INamedTypeSymbol? BaseType => throw new NotImplementedException();
-
-    public ImmutableArray<ISymbol> GetMembers() => [];
-
-    public ImmutableArray<ISymbol> GetMembers(string name) => [];
-
-    public ITypeSymbol? LookupType(string name)
-    {
-        throw new NotImplementedException();
-    }
-}
-
-internal interface IErrorTypeSymbol : INamedTypeSymbol
-{
-}
-
-public class TypeInfo
-{
-    internal TypeInfo(ITypeSymbol type, ITypeSymbol? convertedType)
-    {
-        Type = type;
-        ConvertedType = convertedType;
-    }
-
-    public NullabilityInfo ConvertedNullability { get; }
-
-    public ITypeSymbol? ConvertedType { get; }
-
-    public NullabilityInfo Nullability { get; }
-
-    public ITypeSymbol? Type { get; }
-}
-
-[System.Diagnostics.DebuggerDisplay("{GetDebuggerDisplay(), nq}")]
-public readonly struct NullabilityInfo : IEquatable<NullabilityInfo>
-{
-    public NullableAnnotation Annotation { get; }
-
-    public NullableFlowState FlowState { get; }
-
-    public bool Equals(NullabilityInfo other)
-    {
-        throw new NotImplementedException();
-    }
-}
-
-public enum NullableAnnotation
-{
-    None,
-    NotAnnotated,
-    Annotated
-}
-
-public enum NullableFlowState
-{
-    None,
-    NotNull,
-    MaybeNull
-}
-
-public static class SymbolExtensions
-{
-    /// <summary>
-    /// Unwraps the actual type of a property symbol or local symbol. 
-    /// If the symbol is a type symbol, then return it, otherwise null.
-    /// </summary>
-    /// <param name="symbol"></param>
-    /// <returns></returns>
-    public static ITypeSymbol? UnwrapType(this ISymbol symbol)
-    {
-        if (symbol is IPropertySymbol propertySymbol) return propertySymbol.Type;
-
-        if (symbol is IFieldSymbol fieldSymbol) return fieldSymbol.Type;
-
-        if (symbol is ILocalSymbol localSymbol) return localSymbol.Type;
-
-        return symbol as ITypeSymbol;
     }
 }
