+++ src/Raven.CodeAnalysis/Symbols/IErrorTypeSymbol.cs
@@ -1,0 +1,5 @@
+namespace Raven.CodeAnalysis.Symbols;
+
+internal interface IErrorTypeSymbol : INamedTypeSymbol
+{
+}
