+++ src/Raven.CodeAnalysis/TypeInfo.cs
@@ -1,0 +1,18 @@
+namespace Raven.CodeAnalysis;
+
+public class TypeInfo
+{
+    internal TypeInfo(ITypeSymbol type, ITypeSymbol? convertedType)
+    {
+        Type = type;
+        ConvertedType = convertedType;
+    }
+
+    public NullabilityInfo ConvertedNullability { get; }
+
+    public ITypeSymbol? ConvertedType { get; }
+
+    public NullabilityInfo Nullability { get; }
+
+    public ITypeSymbol? Type { get; }
+}
